canonicalize: target-gated loop<->map interchange, WCR-safe NSDFG expand, perf finalize
Checkpoint of the canonicalization correctness/perf work (TSVC 151/151 throughout).

Interchange + stride (P2):
- NEW MoveLoopIntoMapGated: target-gated loop<->map interchange (GPU always;
  CPU only when it lowers the global-weighted innermost stride), the loop<->map
  stride minimizer the map-only/loop-only passes cannot express.
- Extract score_indexed_strides into minimize_stride_permutation (shared scorer;
  MinimizeStridePermutation delegates, behavior-preserving).
- Pipeline: StateFusionExtended as an early cleaning pass + an 'interchange'
  stage after parallelize (both modes); target threaded through _build_stages.

Reduction-error fix (atax/syr2k/syrk now pass; gramschmidt/correlation no longer
error): make ExpandNestedSDFGInputs WCR-safe on both sides --
- _collect_write_subsets skips WCR out-edges (the 3-tuple upgrade loop already
  skips them, so collecting their 2-tuple crashed the downstream unpack).
- _rewrite_memlets_with_offset preserves memlet.wcr through the offset rewrite
  instead of asserting it away (a reduction only relocates; dropping the WCR
  miscompiled gramschmidt/correlation).

Perf (P1):
- NEW finalize_for_target: post-canon set_fast_implementations + infer_types +
  expand_library_nodes + storage opts (separate from the clean form).

Plus prior verified work folded into this checkpoint: math.h ipow(a,0)=1 (+test),
accumulator/normalize_wcr/split_tasklets refinements, npbench/polybench corpus
loaders, cloudsc/icon snippets loader, speedup_vs_autoopt perf test.

+++ tests/corpus/npbench/dense_linear_algebra/permute_3d.py
@@ -15,7 +15,7 @@ OUTPUT_ARGS = ('B', )
 N = dc.symbol('N', dtype=dc.int64)
 
 
-def initialize(N, datatype=np.float64):
+def initialize(N, datatype=np.float32):
     A = np.fromfunction(lambda i, j, k: (i * N * N + j * N + k * N) / N, (N, N, N), dtype=datatype)
     B = np.zeros((N, N, N), dtype=datatype)
     return (A, B)
